Fix race in nosv_detach
As @arocanon pointed out, when we do worker_free_external, we call sfree which then uses the local CPU slab bucket.
However, if we wake another worker before, we may race at the CPU bucket with the waked up worker.
To prevent this, move the waking up of the next worker after we free the current.