Fix memory ordering race on shm initialization
When using a high number of nOS-V processes, an issue creeps up
due to flock not issuing a memory barrier (because it is not
designed for multithreading).

In this case, it is possible that a process A initializes the shared
memory, and a process B then grabs all the available CPUs.
When B enters the scheduler for the first time, it sees an
uninitialized scheduler->tasks value different from zero,
and proceeds to segfault searching for a task, because
it doesn't see the WRITE from process A to scheduler->tasks.