Fix possible null-dereference in scheduler_get_internal
Although I don't have a specific scenario in mind, it may be
possible for a thread to enter the scheduler and have its
process_scheduler_t be NULL (in scheduler->queues_direct) because
it's only initialized on task insertion (and the process may have
inserted no tasks so far).

Let's preventively fix that one with a null check. We guarantee
that there are tasks inside the scheduler at that point, so at least
one process_scheduler_t is inside the scheduler->queues list.