Fix races in nosv_shutdown
When unregistering a process and its PID from the pidmanager->pids
bitset, it is possible that another process immediately
registers the same logical PID, generating an ABA problem.
In this situation, pidmanager_transfer_to_idle can issue a
cpu_transfer to the same logical_pid, causing worker_wake to
create a new local thread, and hanging the library.

To solve this problem, the best way is to have two phases to free
a logical_pid identifier: first in the pids bitset, and then
in a new pids_alloc bitset. Here we eliminate the ABA problem.

An unrelated issue happenned due to the delegate thread being
joined after the workers. If there is a request in the delegate
queue from a remote process, and we don't process it until all
other threads are joined, we may create a thread after the joining
is done.
Then, this new thread can segfault when the shared memory is
munmaped.
To prevent this, join first the delegate thread, making sure
that no other threads are pending creation when we start
the join process.