Add current worker into idle list before waking up successor
On worker yield, add first the worker into the idle list before waking
up the next worker. When threads are awakened after being blocked for
some time, they usually preempt the waking process. This prevents the
idle thread to add itself to the idle list. This might lead to extra
workers being created if the idle pool is empty.