Remove task_complete from task context
API calls which require a task context cannot be called from the
task_complete callback, since it may be called from external threads.
The best solution is to return EINVAL when this happens, which can be
done by removing the worker from task context as soon as the task_end
call finishes.