Fix nosv_detach affinity instrumentation
During nosv_detach, we need to inform the instrumentation that the
detached thread might overlap with other nosv threads. To do so, we set
it to the cooling state.

Also, inform the instrumentation of the right thread affinity if we are
restoring a previous one. The restored affinity might be translated to
-1 in the case of an affinity with more than one core or a specific cpu
affinity otherwise.