Correctly record execution count for servers
Parallel tasks did not show the correct execution count when the served
task came from the server directly, serving himself at the end of the
execution loop. This happenned because the output parameter was never
updated for that case. Solve the issue by updating the parameter when a
server serves himself.