Change nosv_spin_t implementation to a custom spinlock
The affinity support needs an spinlock that can be initialized
statically. pthread_spin_t needs to be initialized with
pthread_spin_init and nosv_spin_t is a simple wrapper around the pthread
spinlock. This patch modifies nosv_spin_t to use a custom spinlock,
based on the pthread_spin_t glibc implementation, but that can be
initalized statically using the macro NOSV_SPINLOCK_INITIALIZER.

Note that pthread_spin_init simply sets the spinlock to 0, which implies
that static pthread_spin_t are valid even if they are no initialized
with pthread_spin_init. However, for complete compilance, we implement
our own spinlock.