Allow waitfor tasks to be woken up before its timeout expires
This patch enables waitfor tasks to be woken up before its timeout
expires. To do so, re-submit the deadline task using the nosv_submit
flag NOSV_DEADLINE_WAKE.

When the wake operation is issued, the waitfor task might be in
different states: PENDING to be added into the deadline rbtree, WAITING
inside the rbtree, READY to be run inside a queue, or it might be NONE
of the other either because it has not yet called nosv_waitfor or
because it has just exit nosv_waitfor.

The user can be sure than once a NOSV_DEADLINE_WAKE operation is
submitted, the target task will either resume from the nosv_waitfor
regardless of its remaining sleep time or it will prevent it from
sleeping the next time the target task calls nosv_waitfor in case the
target task had not yet called nosv_waitfor.

In simple terms, the effect of NOSV_DEADLINE_WAKE is to change the
deadline state to READY regardless of its previous state. Therfore, the
effect is not cumulative. If a task in the READY state is target of a
NOSV_DEADLINE_WAKE operation, the operation has no effect on the task.

This patch also introduces suprious nosv_waitfor wake ups as a side
effect. This might happen, for instance, when issuing NOSV_DEADLINE_WAKE
on a task that has not yet called nosv_waitfor; the next time the target
tasks calls the api, it will return immeditelly. Also, it can happen
when a thread calls nosv_waitfor in a loop, and the user tries to target
a specific call to nosv_waitfor on iteration "i". The NOSV_DEADLINE_WAKE
might end up targetting the "i + 1" loop instead of the "i" loop,
resulting on an unexpected wake up. Therfore, user code should write
robust code protecting from spurius wake ups on nosv_waitfor.

This patch introduces a new deadline_state_t data type per task with can
held any of the previously mentioned states (PENDING, WAITING, READY,
NONE). The states are used to detect at which point in the scheduler the
deadline task is found, and to move it to the ready queues as soon s
possible. If the task is already WAITING in the red-black tree, the task
is removed from the tree to the ready queues by the thread holding the
dt lock. To do so, any thread can request a "purge" by incrementing the
scheduler->deadline_purge counter. At the same turn, the thread holding
the dtlock will remove as many task as scheduler->deadline_purge and
then decrease the counter by the same number even if less tasks are
found, given that deadline tasks might exit prematurely if their
deadline expires.