Change handling of parallel tasks
Use a new task_execution_handle_t to pass around the runtime which
includes the task execution id, to strengthen the handling of parallel
tasks and splat a couple of bugs.

This also removes the need for the confusing task_stack that was
introduced back in the original implementation of parallel tasks.

Finally, future bugs involving parallel tasks should be easier to fix.