Instrument parallel tasks
Introduces the "body id", a unique identifier for each execution
"instance" of a single parallel task with the same task id. Only
parallel tasks can use a body id > 0.

The ovni_thread_require() function is used to require the version 2.0.0
of the nOS-V emulation model, as it breaks the format of the VT* events.

Co-authored-by: Vincent A. Arcila <vincent.arcila@bsc.es>