Do not break lookup scope order on bypass wrappers
Nosv provides interceptors for functions aimed at faking the affinity of
the running applications. To avoid triggering the fake affinity
mechanism, nosv code can call the bypass_* family of functions to get
the original behaviour of the corresponding function.

However, the bypass_* functions called the next corresponding symbol in
the lookup scope with respect to the nosv library, i.e., it skipped
interceptors that might be defined by other libraries that appear before
nosv library in the lookup scope order. This behaviour triggered an
error in the LLVM implementation of libasan, which could interpose the
pthread_join calls but not the pthread_create calls originating from the
nosv code (because nosv provides an interceptor for pthread_create but
not for pthread_join).

This patch forces bypass_* functions to respect the lookup scope order,
as expected, but use a tls flag to avoid triggering the "fake affinity"
mechanism.

The tls flag is actually a counter, which prevents tiggering the "fake
affinity" mechanism in the case where a a call to a bypass_* function
ends up calling multiple nosv interceptors indirectly. In this scenario,
the counter prevents activating the "fake affinity" mechanism for any of
the nosv interceptors called during the resolution of a single call to a
bypass_* function (this is unlikely to happen, but you never know...).