Fix nested init and shutdown instrumentation
Nosv supports nesting nosv_init and nosv_shutdown calls by using a
refcount. Only the outermost calls will actually init and shutdown the
runtime. In all the other cases, instrumentation events are _always_
emitted. However, if the same thread nests multiple init and shutdown
calls, the resulting sequence of instrumentation events will not make
sense. That is because multiple init and end events are emitted. In
particular, the aplication crashes when a thread attempts to emit the
second "end" event.

This commit introduces a per-thread refcount, which allows a thread to
nest multiple calls to nosv_init/nosv_shutdown and only emit the
instrumentation events on the outermost calls, if necessary.

The outermost calls will only emit the events if the calling thread is
not attached. This is needed to support scenarios in which an attached
thread calls an external library function that attempts to initializes
nosv. nosv_attach'ed threads must have emitted the thread initialization
events prior to calling nosv_attach. Therefore, nosv_init/nosv_shutdown
will not emit these events if an external source already emitted them.

This patch also adapts the corresponding tests to check the new use
cases.