Instrument nosv_attach() and nosv_detach() API
Until now, the instrumentation of the attach/detach API was done by
emitting an attached event at the beginning of nosv_attach() and
another detached event at the end of nosv_detach(). Then, the attached
state was kept in the subsystem view even if no frame was help in the
thread stack.

The subsystem rules enforce that the state must match the thread stack,
so the current implementation is breaking this rule. To correct this
situation, the nosv_attach() and nosv_detach() functions are
instrumented as the other API functions with their respective enter/exit
events, so no "attached" state is held in between.

On the other hand, the subsystem view will show the state "in task body"
between the nosv_attach() and nosv_detach() calls.

The version of the nOS-V instrumentation model is bumped from 1.0.0 to
1.1.0 to include the new VA{aAeE} events. The old "attached" pair of
events VH{aA} are not used anymore.