Introduce coordinated shutdown
This patch introduces the "coordinated shutdown", a mechanism for the
orderly shutdown of nosv workers.

The original shutdown mechanism relies on waking up all idle threads in
parallel while oversubscribing cores. This leds to a slow shutdown in
applications that create a lot of idle threads (such as Fibonacci), due
to threads trying to access shared structures before exiting.

To fix this issue, the coordinated shutdown mechanism wakes idle threads
sequentially, on a per-core basis. i.e. Worker threads that have seen the
shutdown signal attempt to wake a single idle thread on their core before
exiting.

A special core, named the "leader shutdown core", ensures that all idle
threads have been joined to prevent races, similar to how
threadmanager_shutdown did before.

This patch also introduces the "killer task" in order to fix an inherent
problem of this approch: Start the shutdown mechanism of a processes
whose workers are all in (or in transition to) the idle list. This patch
submits a special "killer" task in order to allow a worker of this
process to obtain ownership of a core and start the shutdown procedure.

However, this approach has (at least) two major drawbacks:
 - In the worst case, only one worker will participate in the shutdown
   of all threads.
 - The leader shutdown core might sleep while waiting for threads to
   join, blocking a core from running other processes' tasks.

Sill, on a particular setting, the shutdown of Fibonacci has descended
from 5 minutes to 2 seconds, without an appreciable difference in
shutdown time for applications that do not create that many threads
(tested by timing "make check" with and without this patch).