Optimize delegation thread join
Instead of joining the delegation thread just before starting the
controlled shutdown (and potentially blocking the current core), this
patch defers the join operation (but not the delegation shutdown
message) to the shutdown leader core.

The shutdown delegation core will attempt to join the delegation thread
without blocking, but note that the shutdown leader core will only
attempt to do so if it can't fetch an idle thread to swap first.
Therfore, the leader shutdown core gives time to the kernel to schedule
the delegation thread while it is scheduling idle threads.

Also, only after the delegation thread has been joined, the leader
shutdown core checks if all workers have joined. This is needed to
prevent races where the delegation thread had pending requests to create
new workers queues before the shutdown signal.