Fix heap when size_t is not unsigned long long
Reported-by: Rodrigo Arias Mallo <rodrigo.arias@bsc.es>

When the width of size_t doesn't match the width of the unsigned long
long type, the number of leading zeros doesn't match, making the
heap_get_move() function return incorrect values. This is the case on
ARMv7 with 32 bits, where size_t is 32 bits but unsigned long long is 64
bits.

We can use _Generic to find an appropiate builtin at compile-time which
matches the size of the used type.

See: https://pm.bsc.es/gitlab/rarias/ovni/-/issues/193