Directly swap to ready task when pausing if possible
When pausing a task, nosv always switches to an idle worker, who gets a
task from the scheduler and runs it (possibly switching to another
worker).

Instead, this patch mimics the yield optimization for task pause, where
we first attempt to get a task from the scheduler and switch to it if
possible.