Add acq/release barriers on complete_callbacks
We use acquire/release semantics on user code to prevent reorders
between nOS-V internals and user calls, and in order to ensure that a
complete_callback happens *after* other run and ended callbacks.

Apparently, the NVIDIA Grace chip has a weaker memory model than other
arm64 processors we have tested with. The basic-parallel test was
failing, apparently because a thread did not see the write to task_cnt.
While it's true that seq_cst in atomic_load and atomic_store only
guarantees loads and stores to *that* variable, and offers no guarantees
as to the values seen by other loads and stores to different variables,
I don't recall having seen an architecture that can speculate past these
kind of accesses so as to cause this kind of reorder.

Anyway, what is clear is that we're lacking acquire/release semantics
for the complete callback, in order to ensure that the thread executing
said call, wether in parallel tasks or through the event API can see
every store caused by any other thread during the run_callback and
end_callback (released through their respective atomic barriers).