Move schedpoint instr event to level 4
In order to be able to generate a breakdown trace we need at least the
level 3, but the nosv_schedpoint API event floods the trace when using
the libompv runtime. As most of those events are not required for the
breakdown, we can leave them in the level 4 for when they are needed.
The nOS-V ovni model version is bumped in the patch level so we know
that we have this change in effect from the ovni side.

Fixes: https://gitlab.bsc.es/nos-v/nos-v/-/issues/37