Remove hungry from non-blocking scheduler_get()
When calling scheduler_get() with the non-blocking flag, we are not
going to continue waiting if there is no more work, so we don't need to
mark the thread as hungry. This removes the hungry events for those
cases that we call nosv_yield and nosv_schedpoint, so we can have
smaller traces.