Emit task end before the complete callback
In order to allow non-parallel tasks to be submitted themselves again
from the complete callback, we must guarantee that the task execute
event never occurs prior to emitting the end event. Moving the end event
before the complete callback runs solves the problem, at the cost of
missing the execution of the complete callback in the trace. The
subsystem view can be used to infer that the callback is running for a
long time, as we are not seeing the hungry state.

We bump the patch version number of the nOS-V model so we can determine
if the trace has this patch applied or not.

Fixes: https://gitlab.bsc.es/nos-v/nos-v/-/issues/40
See: https://gitlab.pm.bsc.es/rarias/ovni/-/issues/221