Fix nosv_cond test in NGPP
The condition in the while loop for non-timed tasks has to include
negative ready values.

Otherwise, if a timed tasks arrives to line 75 before one of the
non-timed tasks has finished will execute ready--.
The ready variable will then be a negative value, this check will return
false for non-timed tasks, and they won't go into nosv_cond_wait.
This is the root cause of the failures in tests for NGPP.
If we include negative values as a meaning for waiting, the tests work as designed.